requirements.txt
audio-recorder-streamlit
streamlit
python-dotenv
deepgram-sdk
